https://chronobank.atlassian.net/browse/MD-1275 - fixed bug with env for auth

+++ tests/performance/index.js
@@ -145,9 +145,10 @@ module.exports = (ctx) => {
             return;
 
           const content = JSON.parse(msg.content);
-          expect(content.address).to.equal(address);
-          await ctx.amqp.channel.deleteQueue('test_addr');
-          res();
+          if (content.address == address) { 
+            await ctx.amqp.channel.deleteQueue('test_addr');
+            res();
+          }
         }));
       })()
     ]);
